Refactor service links and update content across multiple pages for improved clarity and navigation
- Updated links to the AI Automation Sprint service to include the correct path.
- Enhanced descriptions in the AI Automation Consulting and AI Integration Services sections for better understanding.
- Revised titles and meta descriptions for various service pages to align with SEO best practices.
- Improved button links and text for consistency and clarity in the user interface.
- Adjusted sitemap entries to reflect the correct URLs and change frequencies.
- Made minor content adjustments for grammatical accuracy and readability across several components.
