+++ app/ai-automation-sprint/page.tsx
@@ -1,1 +1,23 @@
-export { metadata, default } from "../ops-efficiency-sprint/page";
+import { Metadata } from "next";
+import SprintMarketingPage from "../ops-efficiency-sprint/sprint-page";
+
+export const metadata: Metadata = {
+    title: "30-Day Drop-In AI Automation Sprint",
+    description:
+        "In 30 days, get one production-grade AI automation running on top of your existing systems with measurable before/after impact. No migration. No lock-in. No disruption.",
+    keywords: [
+        "30-day AI automation sprint",
+        "enterprise AI automation",
+        "drop-in AI",
+        "workflow automation",
+        "ERP CRM automation",
+    ],
+    alternates: {
+        canonical: "https://www.invaritech.ai/ai-automation-sprint/",
+    },
+};
+
+export default function AIAutomationSprintPage() {
+    return <SprintMarketingPage />;
+}
+
