+++ app/services/enterprise-ai-chatbot-deployment/page.tsx
@@ -2,7 +2,7 @@ import { Metadata } from "next";
 import ChatbotClient from "./chatbot-client";
 
 export const metadata: Metadata = {
-    title: "Enterprise AI Chatbot Deployment | AI Chatbot Development Services",
+    title: "AI Chatbot Development Services | Enterprise Deployment (Governed & Multilingual)",
     description: "AI chatbot development services for enterprises that need governance. Deploy a controlled enterprise AI chatbot solution with RBAC/ABAC, audit logs, citations, tool boundaries, and multilingual support.",
     keywords: [
         "ai chatbot development services",
