+++ components/asterisk-math.tsx
@@ -125,12 +125,12 @@ export default function AsteriskMath() {
                                 projections based on your specific workflow and
                                 team size.
                             </p>
-                            <a
-                                href="/assessment"
-                                className="inline-flex items-center justify-center px-6 py-3 bg-primary text-primary-foreground rounded-lg hover:bg-primary/90 transition-colors"
-                            >
-                                Get your Ops Efficiency Score
-                            </a>
+	                            <a
+	                                href="/assessment/"
+	                                className="inline-flex items-center justify-center px-6 py-3 bg-primary text-primary-foreground rounded-lg hover:bg-primary/90 transition-colors"
+	                            >
+	                                Get your Ops Efficiency Score
+	                            </a>
                         </CardContent>
                     </Card>
                 </div>
